+++ packages/gestalt/src/Pog.js
@@ -22,12 +22,12 @@ type Props = {|
     | 'transparentDarkGray'
     | 'gray'
     | 'lightGray'
-    | 'darkGray'
     | 'white'
     | 'blue',
   dangerouslySetSvgPath?: { __path: string },
   focused?: boolean,
   hovered?: boolean,
+  selected?: boolean,
   iconColor?: 'gray' | 'darkGray' | 'red' | 'blue' | 'white' | 'orange',
   icon?: $Keys<typeof icons>,
   size?: $Keys<typeof SIZE_NAME_TO_PIXEL>,
@@ -50,19 +50,24 @@ export default function Pog(props: Props) {
     dangerouslySetSvgPath,
     focused = false,
     hovered = false,
-    iconColor = defaultIconButtonIconColors[bgColor],
+    iconColor,
     icon,
+    selected = false,
     size = 'md',
   } = props;
 
   const iconSize = SIZE_NAME_TO_PIXEL[size] / 2;
+  const color =
+    (selected && 'white') || iconColor || defaultIconButtonIconColors[bgColor];
 
   const inlineStyle = {
     height: SIZE_NAME_TO_PIXEL[size],
     width: SIZE_NAME_TO_PIXEL[size],
   };
 
-  const classes = classnames(styles.pog, styles[bgColor], {
+  const classes = classnames(styles.pog, {
+    [styles[bgColor]]: !selected,
+    [styles.selected]: selected,
     [styles.active]: active,
     [styles.focused]: focused,
     [styles.hovered]: hovered && !focused && !active,
@@ -79,7 +84,7 @@ export default function Pog(props: Props) {
         */}
         <Icon
           accessibilityLabel=""
-          color={iconColor}
+          color={color}
           dangerouslySetSvgPath={dangerouslySetSvgPath}
           icon={icon}
           size={iconSize}
@@ -113,5 +118,6 @@ Pog.propTypes = {
     'orange',
   ]),
   icon: PropTypes.oneOf(Object.keys(icons)),
+  selected: PropTypes.bool,
   size: PropTypes.oneOf(Object.keys(SIZE_NAME_TO_PIXEL)),
 };